+++ src/components/SavedViewsTable.tsx
@@ -23,7 +23,7 @@ import Tooltip from '@mui/material/Tooltip';
 import Typography from '@mui/material/Typography';
 import { ClusterResolution } from '../lib/clusterIdentity';
 import { findResourceByRouteName } from '../lib/resourceCatalog';
-import { buildSavedViewUrl } from '../lib/savedViewUrl';
+import { buildSavedViewSearchParams, buildSavedViewUrl } from '../lib/savedViewUrl';
 import { SavedView } from '../types';
 
 export interface SavedViewsTableProps {
@@ -160,7 +160,11 @@ export function SavedViewsTable({
                 <Tooltip title={openTooltip}>
                   <span>
                     {canOpen ? (
-                      <Link routeName={view.resource.routeName} params={{ cluster: view.cluster }}>
+                      <Link
+                        routeName={view.resource.routeName}
+                        params={{ cluster: view.cluster }}
+                        search={buildSavedViewSearchParams(view)}
+                      >
                         <IconButton size="small" aria-label={`Open ${view.name}`}>
                           <Icon icon="mdi:open-in-new" width={18} />
                         </IconButton>